Refactor grid area client
